Change base conversion implementation
Changed implementation from using the direct conversion algorithm to for
numbers in base 2, 8 and 16 to doing it in two steps: first convert to
base 10, then to desired destination base for the number.

Also added input tests for all the functions, and changed their
parameters. They now take a char* and a length to avoid using strlen
and other string functions.

+++ include/lbaseconv/bin2int.h
@@ -6,8 +6,9 @@
 */
 #ifndef __BIN2INT_H__
 #define __BIN2INT_H__
+#include <stddef.h>
 
-int btoi (const char *s, int len);
+long btoi (const char *s, size_t len);
 
 #endif /* __BIN2INT_H__ */
